DECIMAL-ACCEPT

+++ AltecSystem.Domain/Entities/User.cs
@@ -7,7 +7,7 @@ namespace AltecSystem.Domain.Entities
         public string Name { get; set; } = string.Empty;
         public string Role { get; set; } = string.Empty;
         public Guid Id { get; set; }
-        public int? Altec_Points { get; set; }
+        public decimal? Altec_Points { get; set; }
 
     }
 }
